Fix compiler warnings (Ubuntu)
git-svn-id: http://root.cern.ch/svn/root/trunk@38409 27541ba8-7e3a-0410-8455-c3a389f83636

+++ cint/cint/inc/DataMbr.h
@@ -53,7 +53,7 @@ G__DataMemberInfo {
   long Offset() ;
   int Bitfield();
   int ArrayDim() ;
-  size_t MaxIndex(int dim) ;
+  long MaxIndex(int dim) ;
   G__ClassInfo* MemberOf();
   void SetGlobalcomp(G__SIGNEDCHAR_T globalcomp);
   int IsValid();
